OpenGL
NOW WE HAVE OPENGL SUPORT, YOU CAN INTERACT WITH THE UI AND YOU EVEN HAVE A WAY TO CHANGE BETWEEN SCENES/GAMES!!!
HotFix: now delta_time is working witht he fps cap :)

+++ src/Iron_Engine/Math/Mathf.hpp
@@ -2,12 +2,12 @@
 class Mathf
 {
 public:
-	static void lerp(double& a, const double& b, const double& t)
+	static void lerp(float& a, const double& b, const double& t)
 	{
 		a += (b - a) * t;
 	}
 
-	static void clamp(double& a, const double& min, const double& max)
+	static void clamp(float& a, const double& min, const double& max)
 	{
 		if (a < min)
 			a = min;
